First Version of the Player App.
Voice/Braille Simulator is not implemented yet.

This version of the player reads from a text file. The text file is
specially formatted to have keywords that will tell the player what to
do. The player right now can separate some of the keywords from the
text. The "repeat" keyword has been implemented.