Second Version of Player
Text/Speech not incorporated yet.

Linked the simulator with the Player.  Added a <checkpoint< function for
the <repeat< function. Added a random scenario generator <random<.
Partially added a <sfx< function. The Player App will read the the <sfx<
keyword and isolate a file name, but will not play the audio.